extracted out common utils
